+++ docs/process/release.md
@@ -25,11 +25,12 @@ Once the release is complete, the assigned release driver will:
     * Cargo.toml should not specify versions for each entry in `dev-dependencies`.
     * Ensure that any new packages have suitable version numbers.
     * Ensure that the Cargo.toml file includes a set of fields consistent with other ICU4X packages.
+* Temporarily add `icu_testdata` back to the workspace
+  * To build, run `cargo make testdata-legacy-gen` to generate the gitignored data
 * `cargo workspaces version --no-git-push --no-git-tag` to bump the version numbers
   * This will only update crates that have changed, and will ask you which version number to bump for each crate
   * You can use commands like `git log icu@1.0.0..@ -- components/plurals/src/` and `cargo public-api -p icu_list diff 1.0.0` to figure out whether to do a major, minor, or patch release
   * Get this reviewed and checked in before continuing
-* `cargo make testdata-legacy` to generate data for `icu_testdata` (which is gitignored)
 * Use `cargo workspaces publish --from-git` to automatically publish the crates in the correct order
   * Add `icu4x-release` group as owners to each new component you're publishing
     * `cargo owner -a github:unicode-org:icu4x-release`
