+++ provider/testdata/Cargo.toml
@@ -33,16 +33,16 @@ include = [
 all-features = true
 
 [dependencies]
-icu_provider = { version = "1.2.0", path = "../core" }
-icu_provider_adapters = { version = "1.2.0", path = "../adapters" }
+icu_provider = { version = "1.2.0", path = "../core", default-features = false }
+icu_provider_adapters = { version = "1.2.0", path = "../adapters", default-features = false }
 
 # buffer feature
-icu_provider_blob = { version = "1.2.0", path = "../blob", optional = true }
+icu_provider_blob = { version = "1.2.0", path = "../blob", default-features = false, optional = true }
 
 # databake deps
-icu_locid = { version = "1.2.0", path = "../../components/locid" }
-icu_collections = { version = "1.2.0", path = "../../components/collections" }
-zerovec = { version = "0.9.4", path = "../../utils/zerovec" }
+icu_locid = { version = "1.2.0", path = "../../components/locid", default-features = false }
+icu_collections = { version = "1.2.0", path = "../../components/collections", default-features = false }
+zerovec = { version = "0.9.4", path = "../../utils/zerovec", default-features = false }
 
 # databake options
 icu_calendar = { version = "1.2.0", path = "../../components/calendar", default-features = false, optional = true }
@@ -62,8 +62,7 @@ icu_segmenter = { version = "1.2.0", path = "../../components/segmenter", defaul
 icu_timezone = { version = "1.2.0", path = "../../components/timezone", default-features = false, optional = true }
 
 [dev-dependencies]
-icu = { path = "../../components/icu" }
-icu_provider = { path = "../../provider/core", features = ["deserialize_json"] }
+icu = { path = "../../components/icu", default-features = false }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
 criterion = "0.4"
